init the rl, works for one link

+++ planarenvs/n_link_reacher/envs/tor.py
@@ -16,15 +16,15 @@ class NLinkTorReacherEnv(NLinkReacherEnv):
         self.observation_space = spaces.Dict(
             {
                 "x": spaces.Box(
-                    low=-self._limUpPos, high=self._limUpPos, dtype=np.float64
+                    low=-self._limUpPos, high=self._limUpPos, dtype=np.float32
                 ),
                 "xdot": spaces.Box(
-                    low=-self._limUpVel, high=self._limUpVel, dtype=np.float64
+                    low=-self._limUpVel, high=self._limUpVel, dtype=np.float32
                 ),
             }
         )
         self.action_space = spaces.Box(
-            low=-self._limUpTor, high=self._limUpTor, dtype=np.float64
+            low=-self._limUpTor, high=self._limUpTor, dtype=np.float32
         )
 
     def continuous_dynamics(self, x, t):
