+++ planarenvs/planar_common/acc_env.py
@@ -7,15 +7,15 @@ class AccEnv(object):
         self.observation_space = spaces.Dict(
             {
                 "x": spaces.Box(
-                    low=-self._limUpPos, high=self._limUpPos, dtype=np.float64
+                    low=-self._limUpPos, high=self._limUpPos, dtype=np.float32
                 ),
                 "xdot": spaces.Box(
-                    low=-self._limUpVel, high=self._limUpVel, dtype=np.float64
+                    low=-self._limUpVel, high=self._limUpVel, dtype=np.float32
                 ),
             }
         )
         self.action_space = spaces.Box(
-            low=-self._limUpAcc, high=self._limUpAcc, dtype=np.float64
+            low=-self._limUpAcc, high=self._limUpAcc, dtype=np.float32
         )
 
     def continuous_dynamics(self, x, t):  # pylint: disable=unused-argument
